refactor: isolate CN indexer read-only source resolution (#926)
